+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-05-31
+## [Unreleased] - 2022-06-01
 
 ### Added
 * CycleDistance: implementation of cycle distance described in https://doi.org/10.3390/app12115506
+* CycleEditDistance: implementation of cycle edit distance described in https://doi.org/10.3390/app12115506
 
 ### Changed